backend/requirements.txt
# FastAPI and server
fastapi>=0.116.1,<1.0.0
starlette>=0.49.1,<1.0.0
uvicorn[standard]==0.27.1
python-multipart==0.0.22

# Redis
redis==5.0.1

# Data validation
pydantic==2.6.1
pydantic-settings==2.1.0

# WebSocket
websockets==12.0

# Utilities
python-dotenv==1.0.1
httpx==0.26.0
PyJWT==2.10.1

# Security: Pin to patched versions (CVE-2026-23949, CVE-2026-24049)
# These MUST be installed to override any transitive dependencies
wheel>=0.46.2
jaraco.context>=6.1.0
setuptools>=75.8.0
